finetune/pyproject.toml
[project]
name = "spatial-ai-finetune"
version = "0.1.0"
description = "Model finetuning with Axolotl"
requires-python = ">=3.11,<3.13"
dependencies = [
    "axolotl<0.14.0",
    "pyyaml>=6.0.3",
]
